backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib
python-dotenv
slowapi
pytest
httpx
